Split out AI Chat Lab views into separate directory
