+++ apps/src/aiComponentLibrary/infoTooltipIcon/info-tooltip-icon.module.scss
