+++ apps/src/aichat/types/toxicity.ts
@@ -1,7 +1,6 @@
+import {FIELDS_CHECKED_FOR_TOXICITY} from '@cdo/apps/aichatLab/views/modelCustomization/constants';
 import {ValueOf} from '@cdo/apps/types/utils';
 import {AiChatTeacherFeedback} from '@cdo/generated-scripts/sharedConstants';
-
-import {FIELDS_CHECKED_FOR_TOXICITY} from '../views/modelCustomization/constants';
 
 /** Response structure for the detect toxicity API */
 export interface DetectToxicityResponse {
