+++ apps/src/aichatLab/views/aichatView.module.scss
@@ -1,5 +1,5 @@
 @import 'color.scss';
-@import './aichat-common.module.scss';
+@import '../../aichat/views/aichat-common.module.scss';
 @import '../../mixins.scss';
 
 //200px is a value got from visual comparison (screen collapses when columns width is smaller than 196.7px,
