+++ apps/src/aichatLab/views/model-customization-workspace.module.scss
@@ -1,6 +1,6 @@
 @import 'color';
 @import '../../mixins.scss';
-@import './aichat-common.module.scss';
+@import '../../aichat/views/aichat-common.module.scss';
 
 // Constants
 $default-spacing: 4px;
