+++ apps/src/aichatLab/views/modelCustomization/ExampleTopicsInputs.tsx
@@ -1,10 +1,9 @@
 import Alert, {alertTypes} from '@code-dot-org/component-library/alert';
 import React, {useCallback, useMemo} from 'react';
 
+import {setModelCardProperty} from '@cdo/apps/aichat/redux';
+import {Visibility} from '@cdo/apps/aichat/types';
 import {useAppDispatch} from '@cdo/apps/util/reduxHooks';
-
-import {setModelCardProperty} from '../../redux';
-import {Visibility} from '../../types';
 
 import MultiInputCustomization from './MultiInputCustomization';
