+++ apps/src/aichatLab/views/presentation/model-card-row.module.scss
