+++ apps/test/unit/aichat/utilsTest.ts
@@ -1,6 +1,6 @@
 import * as utils from '@cdo/apps/aichat/redux/utils';
 import {AiCustomizations} from '@cdo/apps/aichat/types';
-import {EMPTY_MODEL_CARD_INFO} from '@cdo/apps/aichat/views/modelCustomization/constants';
+import {EMPTY_MODEL_CARD_INFO} from '@cdo/apps/aichatLab/views/modelCustomization/constants';
 import {AiChatModelIds} from '@cdo/generated-scripts/sharedConstants';
 
 describe('aichatRedux utils', () => {
